[profiler] Introduce runtime control API.

Commit migrated from https://github.com/mono/mono/commit/9f2c4f328db33b7042a04dffd25784ece5bf40f0

+++ src/mono/mono/metadata/profiler-private.h
@@ -49,7 +49,7 @@ typedef struct {
 	MonoProfilerHandle sampling_owner;
 	MonoSemType sampling_semaphore;
 	MonoProfilerSampleMode sample_mode;
-	uint64_t sample_freq;
+	guint32 sample_freq;
 	gboolean allocations;
 
 #define _MONO_PROFILER_EVENT(name) \
